requirements.txt
torch==2.1.2
torchvision
opencv-python
scikit-learn
pandas
numpy==1.26.0
pillow
dill
python-box
python-dotenv
python-multipart

-e .